test: separate e2e test
