+++ qdrant/tests/test_e2e.py
@@ -1,12 +1,14 @@
+
 import pytest
 
-from datadog_checks.base.constants import ServiceCheck
-from datadog_checks.dev.utils import get_metadata_metrics
 from datadog_checks.qdrant import QdrantCheck
+
+from datadog_checks.dev.utils import get_metadata_metrics
+from datadog_checks.base.constants import ServiceCheck
 from tests.common import OM_METRICS
 
 
-@pytest.mark.unit
+@pytest.mark.e2e
 def test_emits_metrics(dd_run_check, aggregator, instance):
     check = QdrantCheck("qdrant", {}, [instance])
     dd_run_check(check)
@@ -19,12 +21,3 @@ def test_emits_metrics(dd_run_check, aggregator, instance):
     aggregator.assert_service_check("qdrant.openmetrics.health", ServiceCheck.OK)
     aggregator.assert_service_check("qdrant.readyz.status", ServiceCheck.OK)
     aggregator.assert_service_check("qdrant.livez.status", ServiceCheck.OK)
-
-
-def test_empty_instance(dd_run_check):
-    with pytest.raises(
-        Exception,
-        match="InstanceConfig`:\nopenmetrics_endpoint\n  Field required",
-    ):
-        check = QdrantCheck("qdrant", {}, [{}])
-        dd_run_check(check)
